% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BS_Option.R
\name{BS_Option}
\alias{BS_Option}
\title{Calculate the value of a given European Option with Black & Scholes formulae as well as some Greeks for further information}
\usage{
BS_Option(
  Call = TRUE,
  S = 100,
  T = 1,
  t = 0,
  K = 75,
  r = 1,
  d = 0,
  sigma = 40,
  ratio = 1
)
}
\arguments{
\item{Call}{Boolean; indicate if the option is an European Call or Put}

\item{S}{Numeric; Current stock value}

\item{T}{Numeric; Time to maturity}

\item{t}{Numeric; Current time (has to be between 0 and T)}

\item{K}{Numeric; Strike Price of the underlying option}

\item{r}{Numeric; Riskfree rate of interest}

\item{d}{Numeric; Annual dividend yield}

\item{sigma}{Numeric; Implicite Volatility of the underlying stock}

\item{ratio}{Numeric; Ratio of the option}
}
\value{
Value of an European Call or Put Option, calculated by Black & Scholes formulae
}
\description{
Calculate the value of a given European Option with Black & Scholes formulae as well as some Greeks for further information
}
