% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockeystick_plot.R
\name{hockeystick_plot}
\alias{hockeystick_plot}
\title{Plot the payoff function of an European Option}
\usage{
hockeystick_plot(Call, T, t, K, r, d, sigma)
}
\arguments{
\item{Call}{Boolean; indicate if the option is an European Call or Put}

\item{T}{Numeric; Time to maturity}

\item{t}{Numeric; Current time (has to be between 0 and T)}

\item{K}{Numeric; Strike Price of the underlying option}

\item{r}{Numeric; Riskfree rate of interest}

\item{d}{Numeric; Annual dividend yield}

\item{sigma}{Numeric; Implicite Volatility of the underlying stock}
}
\value{
Schematic graphic of an Option Payoff Function
}
\description{
Plot the payoff function of an European Option
}
